# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Insurance Quotes API
  description: "Multi-step insurance quote flow: create, coverage, submit."
  version: 1.0.0
servers:
- url: http://localhost
  description: Generated server url
security:
- bearerAuth: []
paths:
  /quotes:
    get:
      tags:
      - quote-controller
      operationId: listQuotes
      parameters:
      - name: API-Version
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          default: 20
          minimum: 1
          maximum: 100
      - name: search
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
          enum:
          - DRAFT
          - SUBMITTED
          - SUBMISSION_FAILED
          - EXPIRED
      - name: coverage
        in: query
        schema:
          type: string
          enum:
          - BASIC
          - STANDARD
          - PREMIUM
      - name: sortBy
        in: query
        schema:
          type: string
          default: createdAt
          enum:
          - createdAt
          - updatedAt
          - name
          - monthlyPremium
          - status
      - name: direction
        in: query
        schema:
          type: string
          default: desc
          enum:
          - asc
          - desc
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuotePage"
    post:
      tags:
      - quote-controller
      operationId: create
      parameters:
      - name: API-Version
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuoteDetails"
  /quotes/{id}/submit:
    post:
      tags:
      - submission-controller
      operationId: submit
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: API-Version
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuoteDetails"
  /quotes/summary:
    get:
      tags:
      - quote-controller
      operationId: getSummary
      parameters:
      - name: API-Version
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuoteSummary"
  /auth/webauthn/register:
    post:
      tags:
      - auth-controller
      operationId: register
      parameters:
      - name: API-Version
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebAuthnRegisterRequest"
        required: true
      responses:
        "204":
          description: No Content
  /auth/webauthn/register-options:
    post:
      tags:
      - auth-controller
      operationId: registerOptions
      parameters:
      - name: API-Version
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WebAuthnChallenge"
  /auth/webauthn/assertion-options:
    post:
      tags:
      - auth-controller
      operationId: assertionOptions
      parameters:
      - name: API-Version
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssertionOptionsRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WebAuthnChallenge"
  /auth/webauthn/assert:
    post:
      tags:
      - auth-controller
      operationId: assertPasskey
      parameters:
      - name: API-Version
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebAuthnAssertRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenPair"
  /auth/refresh:
    post:
      tags:
      - auth-controller
      operationId: refresh
      parameters:
      - name: API-Version
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenPair"
  /auth/logout:
    post:
      tags:
      - auth-controller
      operationId: logout
      parameters:
      - name: API-Version
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "204":
          description: No Content
  /auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      parameters:
      - name: API-Version
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResult"
  /quotes/{id}/coverage:
    patch:
      tags:
      - quote-controller
      operationId: updateCoverage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: API-Version
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCoverageRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuoteDetails"
  /quotes/{id}:
    get:
      tags:
      - quote-controller
      operationId: getQuote
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: API-Version
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/QuoteDetails"
components:
  schemas:
    CreateQuoteRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        age:
          type: integer
          format: int32
          maximum: 120
          minimum: 18
        zipCode:
          type: string
          minLength: 1
          pattern: "\\d{5}"
      required:
      - age
      - email
      - name
      - zipCode
    QuotePage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/QuoteDetails"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
      required:
      - content
      - page
      - size
      - totalElements
      - totalPages
      - hasNext
      - hasPrevious
    QuoteDistribution:
      type: object
      properties:
        key:
          type: string
        count:
          type: integer
          format: int64
      required:
      - key
      - count
    QuoteTrendPoint:
      type: object
      properties:
        date:
          type: string
          format: date
        created:
          type: integer
          format: int64
        submitted:
          type: integer
          format: int64
        failed:
          type: integer
          format: int64
      required:
      - date
      - created
      - submitted
      - failed
    QuoteSummary:
      type: object
      properties:
        totalQuotes:
          type: integer
          format: int64
        draftQuotes:
          type: integer
          format: int64
        submittedQuotes:
          type: integer
          format: int64
        submissionFailedQuotes:
          type: integer
          format: int64
        expiredQuotes:
          type: integer
          format: int64
        pricedQuotes:
          type: integer
          format: int64
        totalMonthlyPremium:
          type: number
        averageMonthlyPremium:
          type: number
        submissionRate:
          type: number
        statusDistribution:
          type: array
          items:
            $ref: "#/components/schemas/QuoteDistribution"
        coverageDistribution:
          type: array
          items:
            $ref: "#/components/schemas/QuoteDistribution"
        trend:
          type: array
          items:
            $ref: "#/components/schemas/QuoteTrendPoint"
      required:
      - totalQuotes
      - draftQuotes
      - submittedQuotes
      - submissionFailedQuotes
      - expiredQuotes
      - pricedQuotes
      - totalMonthlyPremium
      - averageMonthlyPremium
      - submissionRate
      - statusDistribution
      - coverageDistribution
      - trend
    QuoteDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        age:
          type: integer
          format: int32
        zipCode:
          type: string
        coverageType:
          type: string
          enum:
          - BASIC
          - STANDARD
          - PREMIUM
        hasPreexistingConditions:
          type: boolean
        conditions:
          type: array
          items:
            type: string
            enum:
            - DIABETES
            - HEART_DISEASE
            - HYPERTENSION
            - CANCER_HISTORY
            - OTHER
          uniqueItems: true
        takesPrescriptionMedication:
          type: boolean
        usesTobacco:
          type: boolean
        needsSpouseCoverage:
          type: boolean
        monthlyPremium:
          type: number
        status:
          type: string
          enum:
          - DRAFT
          - SUBMITTED
          - SUBMISSION_FAILED
          - EXPIRED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WebAuthnRegisterRequest:
      type: object
      properties:
        challengeId:
          type: string
          minLength: 1
        credentialJson:
          type: string
          minLength: 1
      required:
      - challengeId
      - credentialJson
    WebAuthnChallenge:
      type: object
      properties:
        challengeId:
          type: string
        publicKeyOptionsJson:
          type: string
    AssertionOptionsRequest:
      type: object
      properties:
        username:
          type: string
    WebAuthnAssertRequest:
      type: object
      properties:
        challengeId:
          type: string
          minLength: 1
        credentialJson:
          type: string
          minLength: 1
        mfaToken:
          type: string
      required:
      - challengeId
      - credentialJson
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresInSeconds:
          type: integer
          format: int64
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 1
      required:
      - refreshToken
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
      - password
      - username
    LoginResult:
      type: object
      properties:
        status:
          type: string
        mfaToken:
          type: string
        tokens:
          $ref: "#/components/schemas/TokenPair"
    UpdateCoverageRequest:
      type: object
      properties:
        coverageType:
          type: string
          enum:
          - BASIC
          - STANDARD
          - PREMIUM
        hasPreexistingConditions:
          type: boolean
        conditions:
          type: array
          items:
            type: string
            enum:
            - DIABETES
            - HEART_DISEASE
            - HYPERTENSION
            - CANCER_HISTORY
            - OTHER
          uniqueItems: true
        takesPrescriptionMedication:
          type: boolean
        usesTobacco:
          type: boolean
        needsSpouseCoverage:
          type: boolean
      required:
      - coverageType
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT